fix(discovery): close recovered lifecycle gaps (#74)
