+++ crates/app/testing/peritus-bug-discovery/src/bin/ndjson.rs
@@ -1,8 +1,5 @@
 //! Bounded raw NDJSON target.
 #![no_main]
 
-libfuzzer_sys::fuzz_target!(|bytes: &[u8]| {
-    if bytes.len() <= peritus_bug_discovery::MAX_INPUT_BYTES {
-        peritus_bug_discovery::check_input(peritus_bug_discovery::DiscoveryTarget::Ndjson, bytes);
-    }
-});
+#[path = "ndjson/fuzz_entry.rs"]
+mod fuzz_entry;
