+++ crates/app/testing/peritus-bug-discovery/src/bin/sse.rs
@@ -1,8 +1,5 @@
 //! Bounded raw SSE target.
 #![no_main]
 
-libfuzzer_sys::fuzz_target!(|bytes: &[u8]| {
-    if bytes.len() <= peritus_bug_discovery::MAX_INPUT_BYTES {
-        peritus_bug_discovery::check_input(peritus_bug_discovery::DiscoveryTarget::Sse, bytes);
-    }
-});
+#[path = "sse/fuzz_entry.rs"]
+mod fuzz_entry;
