+++ crates/app/testing/peritus-bug-discovery/src/bin/working_state.rs
@@ -1,11 +1,5 @@
 //! Bounded structured working-state target.
 #![no_main]
 
-libfuzzer_sys::fuzz_target!(|bytes: &[u8]| {
-    if bytes.len() <= peritus_bug_discovery::MAX_INPUT_BYTES {
-        peritus_bug_discovery::check_input(
-            peritus_bug_discovery::DiscoveryTarget::WorkingState,
-            bytes,
-        );
-    }
-});
+#[path = "working_state/fuzz_entry.rs"]
+mod fuzz_entry;
