+++ xtask/src/cli/help.rs
@@ -53,10 +53,12 @@ Commands:
   distro-upload          Upload a verified package set to the exact existing release draft
   distro-test            Run distribution package tooling regression tests
   discovery-replay       Replay every registered discovery corpus on stable Rust
+  discovery-posix-lifecycle Qualify strict disposable POSIX uninstall failure handling
   discovery-setup-fuzz    Install exact harness-only nightly and cargo-fuzz pins
   discovery-setup-mutation Install the exact cargo-mutants pin
   discovery-mutation-SLICE-N Run fixed mutation shard N (0 through 7)
-  discovery-mutation-context Inventory and test required-context mutations
+  discovery-mutation-context Inventory required-context mutations; zero reachability fails
+  discovery-mutation-context-canary Prove the required-context capacity test rejects its curated mutant
   discovery-mutation-receipt Inventory and test effect-receipt mutations
   discovery-mutation-cancellation Inventory and test product lifecycle mutations
   discovery-fuzz-sse      Run bounded SSE fuzzing and retain completion evidence
